// _js/snippets/product-options/app/modules/smp-sibling-links/style.module.scss.d.ts
/* DO NOT EDIT: This file is auto-generated and will be overwritten. [build-fingerprint:codegen] */
export type Styles = {
  "SmpSiblingLinks": "SmpSiblingLinks",
  "SmpSiblingLinks-badge": "SmpSiblingLinks-badge",
  "SmpSiblingLinks-indicator": "SmpSiblingLinks-indicator",
  "SmpSiblingLinks-indicator--useSwatchImageSize": "SmpSiblingLinks-indicator--useSwatchImageSize",
  "SmpSiblingLinks-indicatorImage": "SmpSiblingLinks-indicatorImage",
  "SmpSiblingLinks-item": "SmpSiblingLinks-item",
  "SmpSiblingLinks-item--current": "SmpSiblingLinks-item--current",
  "SmpSiblingLinks-item--outOfStock": "SmpSiblingLinks-item--outOfStock",
  "SmpSiblingLinks-item--unavailable": "SmpSiblingLinks-item--unavailable",
  "SmpSiblingLinks-label": "SmpSiblingLinks-label",
  "SmpSiblingLinks-list": "SmpSiblingLinks-list",
  "SmpSiblingLinks-title": "SmpSiblingLinks-title",
  "SmpSiblingLinks-titleDescription": "SmpSiblingLinks-titleDescription",
  "SmpSiblingLinks-titleSelection": "SmpSiblingLinks-titleSelection"
}

export type ClassNames = keyof Styles

declare const styles: Styles

export default styles